move graph rewrite to ops [run_proess_replay]
